refactor: Claude request/response dto 클래스로 분리
각 enum들은 적절하게 클래스에 응집
