+++ src/main/kotlin/dev/jxmen/cs/ai/interviewer/adapter/output/external/ClaudeMessage.kt
@@ -1,0 +1,17 @@
+package dev.jxmen.cs.ai.interviewer.adapter.output.external
+
+import com.fasterxml.jackson.annotation.JsonValue
+
+data class ClaudeMessage(
+    val role: ClaudeMessageRole,
+    val content: String,
+)
+
+enum class ClaudeMessageRole {
+    USER,
+    ASSISTANT,
+    ;
+
+    @JsonValue
+    fun toLower(): String = name.lowercase()
+}
